[Faculties] Extract component UpdateButton
This is also needed in StudyProgramme
